+++ include/protozero/byteswap.hpp
@@ -10,6 +10,7 @@ documentation.
 
 *****************************************************************************/
 
+#include <cstdint>
 #include <cassert>
 
 namespace protozero {
@@ -26,14 +27,21 @@ inline void byteswap<1>(const char* data, char* result) {
 
 template <>
 inline void byteswap<4>(const char* data, char* result) {
+# if defined(__GNUC__) || defined(__clang__)
+    *reinterpret_cast<uint32_t*>(result) = __builtin_bswap32(*reinterpret_cast<const uint32_t*>(data));
+# else
     result[3] = data[0];
     result[2] = data[1];
     result[1] = data[2];
     result[0] = data[3];
+#endif
 }
 
 template <>
 inline void byteswap<8>(const char* data, char* result) {
+# if defined(__GNUC__) || defined(__clang__)
+    *reinterpret_cast<uint64_t*>(result) = __builtin_bswap64(*reinterpret_cast<const uint64_t*>(data));
+# else
     result[7] = data[0];
     result[6] = data[1];
     result[5] = data[2];
@@ -42,6 +50,7 @@ inline void byteswap<8>(const char* data, char* result) {
     result[2] = data[5];
     result[1] = data[6];
     result[0] = data[7];
+#endif
 }
 
 } // end namespace protozero